submissions/backend-senior/Maedeh-Shahabi/escrow_account/product/serializers/image_serializer.py
from rest_framework import serializers

from product.models import ImageModel


class ImageSerializer(serializers.ModelSerializer):
    """
    A serializer for our image model
    """
    class Meta:
        model = ImageModel
        fields = [
            'image',
        ]
