submissions/backend-senior/Maedeh-Shahabi/escrow_account/requirements.txt
asgiref==3.5.2
Django==3.2.13
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.0
Pillow==9.1.1
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
